First Attempt to Simulate PDEs on Graphs
This is a first version for a simulator of PDEs over graphs using simple
direct integration.

It is needed the basic Rioko plug-in in Eclipse to run it, because it
uses the basic Graph structure given in the grapht package.

It allows with a main to create a random graph with some density and
some size and start a heat pde over it using the first node as emisor of
energy with a sin function.
It saves the results obtained in a csv file, which is easily converted
into graphics that allows to watch the real evolution of the energy on
the graph.