Added some Functions and a Wave PDE
Added two new kinds of Functions, that allow to create more specific
examples to try the simulator.

It also improves the Simulation allowing to declare the ellapsed time
for the simulation and taking care of the parameters given.
The Simulator now allows to declare a bigger Boundary Set of vertices on
the graph.

The Simulator class methods have been splitted more so now it is easier
to find where make changes.