Bug fixed and rellocation
A bug on EDOFunction has been fixed and now it calculate the values
correctly.
Before, an auto-referenced function coul call "evaluate" method
repetidely, which cause that variables in the original method crashed.
Now it is independent of the caller of the method, evaluate just extend
the interval of calculated function.

A tests folder has been created, and the project shoul execute in
"${workspace_loc:Simulator}/tests" to work well.
Subfolders were created to help another people to read the results
obtained.